# improve JS build speed