--- a/package.json
+++ b/package.json
@@ -21,10 +21,10 @@
     "babel-jest": "^23.0.1",
     "babel-loader": "^7.1.4",
     "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-react-html-attrs": "^2.0.0",
+    "babel-plugin-react-html-attrs": "^2.1.0",
     "babel-plugin-transform-async-to-generator": "^6.16.0",
     "babel-plugin-transform-class-properties": "^6.18.0",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-decorators-legacy": "^1.3.5",
     "babel-polyfill": "^6.26.0",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
@@ -71,7 +71,8 @@
     "redux-thunk": "^2.1.0",
     "smoothscroll-polyfill": "^0.3.6",
     "uswds": "^1.4.0",
-    "webpack": "^3.6.0"
+    "webpack": "^4.12.0",
+    "webpack-cli": "^3.0.4"
   },
   "jest": {
     "coverageDirectory": "./coverage/",