--- a/package.json
+++ b/package.json
@@ -24,6 +24,7 @@
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
     "babel-preset-stage-2": "^6.24.1",
+    "cache-loader": "^1.2.2",
     "codecov": "^2.3.0",
     "deepmerge": "^1.5.2",
     "dotenv": "^4.0.0",