--- a/package.json
+++ b/package.json
@@ -29,7 +29,6 @@
     "deepmerge": "^1.5.2",
     "dotenv": "^4.0.0",
     "enzyme": "^2.9.1",
-    "es6-promise": "^4.1.1",
     "esdoc": "^1.0.3",
     "eslint": "^4.7.2",
     "eslint-config-standard": "^10.2.1",