This pull request does a number of things to improve the JavaScript build speed:

* Upgrades Webpack and Babel
* Removes plugins that were unused or are now included by default
* Caches the intermediate compiled JS files for faster rebuilds
* Switches to a different kind of source map
* Reduces Webpack's module search space

As a result, **this takes initial build time from 116s to 44s, and subsequent build time from 116s to 19s** 🚀 Bonus: the Webpack config is now cleaner, too ✨