@line47 2018-06-13
I'm getting an error when running `make run`

```
Invalid configuration object. Webpack has been initialised using a configuration object that does not match the API schema.
js_1    |  - configuration has an unknown property 'mode'. These properties are valid:
js_1    |    object { amd?, bail?, cache?, context?, dependencies?, devServer?, devtool?, entry, externals?, loader?, module?, name?, node?, output?, parallelism?, performance?, plugins?, profile?, recordsInputPath?, recordsOutputPath?, recordsPath?, resolve?, resolveLoader?, stats?, target?, watch?, watchOptions? }
js_1    |    For typos: please correct them.
js_1    |    For loader options: webpack 2 no longer allows custom properties in configuration.
js_1    |      Loaders should be updated to allow passing options via loader options in module.rules.
js_1    |      Until loaders are updated one can use the LoaderOptionsPlugin to pass these options to the loader:
js_1    |      plugins: [
js_1    |        new webpack.LoaderOptionsPlugin({
js_1    |          // test: /\.xxx$/, // may apply this only for some modules
js_1    |          options: {
js_1    |            mode: ...
js_1    |          }
js_1    |        })
js_1    |      ]
js_1    | error Command failed with exit code 1.
js_1    | info Visit https://yarnpkg.com/en/docs/cli/run for documentation about this command.
```