@afeld 2018-06-13
Hmm, I was getting the same issue - seems to be a weird Docker caching thing. Try

```sh
docker-compose down && make run
```